refactor: move credential ui to separate file
