+++ packages/cardano-metamask-snap/src/api/cardano__signTransaction/ui/certificate.ts
@@ -7,26 +7,7 @@ import {
 } from '../../derivationPath';
 import { getUiAccountIndex, section } from '../../ui';
 import { ADA_TICKER, assetValue } from './utils';
-
-type BaseTxCredential =
-  | {
-      type: 'keyHash';
-      keyHashBech32: string;
-    }
-  | {
-      type: 'scriptHash';
-      scriptHashBech32: string;
-    };
-
-export type OwnTxCredential<TDerivationPath extends CardanoDerivationPath> =
-  BaseTxCredential & {
-    type: 'keyHash';
-    derivationPath: TDerivationPath;
-  };
-
-export type TxCredential<TDerivationPath extends CardanoDerivationPath> =
-  | BaseTxCredential
-  | OwnTxCredential<TDerivationPath>;
+import { TxCredential, renderCredential } from './credential';
 
 export type DRep =
   | BaseTxCredential
@@ -66,19 +47,6 @@ export type Certificate =
       credential: TxCredential<CardanoStakeDerivationPath>;
       dRep: DRep;
     };
-
-const renderCredential = (credential: TxCredential<CardanoDerivationPath>) => {
-  if (credential.type === 'keyHash') {
-    if ('derivationPath' in credential) {
-      return [
-        row('For key', text(credential.keyHashBech32)),
-        row('Derivation path', text(`${credential.derivationPath.join('/')}`)),
-      ];
-    }
-    return [row('For key', text(credential.keyHashBech32))];
-  }
-  return [row('For script', text(credential.scriptHashBech32))];
-};
 
 const renderAccountIndex = (certificate: Certificate) => {
   if ('derivationPath' in certificate.credential) {
